fix: merge conflicts with provider translation
