+++ mpqp/translation/braket.py
@@ -1,17 +1,17 @@
 from typing import TYPE_CHECKING
 
+from mpqp.core.instruction.gates.gate_decomposition import resolve_gate
+from mpqp.core.instruction.gates.gate import Gate
 from mpqp.environment.var_cache import (
     _INSTALLED_MPQP_PROVIDERS,  # pyright: ignore[reportPrivateUsage]
     InstalledProviders,
 )
 
 if InstalledProviders.BRAKET in _INSTALLED_MPQP_PROVIDERS:
-    if TYPE_CHECKING:
-        from braket.circuits import Circuit as braket_Circuit
-        from mpqp.core.instruction.gates.native_gates import NativeGate
-        from mpqp.core.circuit import QCircuit
+    from braket.circuits import Circuit as braket_Circuit
+    from mpqp.core.circuit import QCircuit
 
-    def braket_to_mpqp(qcircuit: "braket_Circuit") -> "QCircuit":
+    def braket_to_mpqp(qcircuit: braket_Circuit) -> QCircuit:
 
         from braket.circuits.serialization import IRType
         from braket.ir.openqasm.program_v1 import Program
@@ -46,12 +46,26 @@ if InstalledProviders.BRAKET in _INSTALLED_MPQP_PROVIDERS:
             qc.add(noises)
         return qc
 
+    def get_braket_gate_set() -> set[type[Gate]]:
+        """Return gates directly representable by Braket."""
+        from mpqp.gates import CNOT, PRX, Rxx, Ryy, Rzz, Rx, Ry, Rz
+
+        return {
+                Rx,
+                Ry,
+                Rz,
+                PRX,
+                Rxx,
+                Ryy,
+                Rzz,
+                CNOT,
+            }
+
     def mpqp_to_braket(
-        circuit: "QCircuit",
+        circuit: QCircuit,
         skip_pre_measure: bool = False,
         skip_measurements: bool = False,
-        authorized_gates: set[type["NativeGate"]] | None = None,
-    ) -> "braket_Circuit":
+    ) -> braket_Circuit:
         """Translate a MPQP circuit to a Braket equivalent.
 
         Note:
@@ -61,7 +75,6 @@ if InstalledProviders.BRAKET in _INSTALLED_MPQP_PROVIDERS:
             circuit: The original MPQP circuit to be translated.
             skip_pre_measure: If set at True will translate the circuit without its pre-measurement circuit (see QCircuit.to_other_language for more information).
             skip_measurements: If set at True will translate the circuit without any measurement.
-            authorized_gates: The set of gates allowed on the circuit, if the circuit contains any other gates it raises a ValueError.
 
         Examples:
             >>> circuit = QCircuit([H(0), CNOT(0, 1), BasisMeasure()])
@@ -103,17 +116,14 @@ if InstalledProviders.BRAKET in _INSTALLED_MPQP_PROVIDERS:
         )
         from mpqp.core.circuit import QCircuit
 
-        if authorized_gates is None:
-            authorized_gates = set()
         if len(circuit.noises) != 0:
             if any(isinstance(instr, CRk) for instr in circuit.instructions):
                 raise NotImplementedError(
                     "Cannot simulate noisy circuit with CRk gate due to "
                     "an error on AWS Braket side."
                 )
-        from braket.circuits import Circuit as BracketCircuit
 
-        braket_circuit = BracketCircuit()
+        braket_circuit = braket_Circuit()
 
         # If the number of qubits are defined by the user, we ensure that every qubits are used.
         # Otherwise the circuit can remain non continuous.
@@ -154,13 +164,14 @@ if InstalledProviders.BRAKET in _INSTALLED_MPQP_PROVIDERS:
                         braket_circuit.measure(targets)
                 continue
             if isinstance(instruction, Gate):
-                from mpqp.translation.utils import verify_convert_instructions
-
-                instr = verify_convert_instructions(instruction, authorized_gates)
+                instructions = resolve_gate(
+                    instruction,
+                    get_braket_gate_set(),
+                )
             else:
-                instr = [instruction]
+                instructions = (instruction,)
 
-            for instruction in instr:
+            for instruction in instructions:
                 braket_instr = instruction.to_other_language(Language.BRAKET)
                 try:
                     targets = [target for target in instruction.targets]
