+++ mpqp/translation/qiskit.py
@@ -1,14 +1,16 @@
 from typing import TYPE_CHECKING
+
+from mpqp.core.instruction.gates.gate_decomposition import resolve_gate
+from mpqp.core.instruction.gates.gate import Gate
+from mpqp.core.instruction.gates.native_gates import U
 from mpqp.environment.var_cache import (
     _INSTALLED_MPQP_PROVIDERS,  # pyright: ignore[reportPrivateUsage]
     InstalledProviders,
 )
 
 if InstalledProviders.QISKIT in _INSTALLED_MPQP_PROVIDERS:
-    if TYPE_CHECKING:
-        from mpqp.core.instruction.gates.native_gates import NativeGate
-        from mpqp.core.circuit import QCircuit
-        from qiskit import QuantumCircuit
+    from mpqp.core.circuit import QCircuit
+    from qiskit import QuantumCircuit
 
     def qiskit_to_mpqp(qcircuit: "QuantumCircuit"):
         """Translate a qiskit QuantumCircuit into a MPQP QCircuit.
@@ -30,13 +32,17 @@ if InstalledProviders.QISKIT in _INSTALLED_MPQP_PROVIDERS:
         qc = qasm2_parse(qasm2_code)
         return qc
 
+    def get_qiskit_gate_set() -> set[type[Gate]]:
+        from mpqp.gates import CNOT, PRX, Rxx, Ryy, Rzz, Rx, Ry, Rz
+
+        return {Rx, Ry, Rz, PRX, Rxx, Ryy, Rzz, U, CNOT}
+
     def mpqp_to_qiskit(
-        circuit: "QCircuit",
+        circuit: QCircuit,
         skip_pre_measure: bool = False,
         skip_measurements: bool = False,
         printing: bool = False,
-        authorized_gates: set[type["NativeGate"]] | None = None,
-    ) -> "QuantumCircuit":
+    ) -> QuantumCircuit:
         """Translate a MPQP circuit to a Qiskit equivalent.
 
         Note:
@@ -46,7 +52,6 @@ if InstalledProviders.QISKIT in _INSTALLED_MPQP_PROVIDERS:
             circuit: The original MPQP circuit to be translated.
             skip_pre_measure: If set at True will translate the circuit without its pre-measurement circuit (see QCircuit.to_other_language for more information).
             skip_measurements: If set at True will translate the circuit without any measurement.
-            authorized_gates: The set of gates allowed on the circuit, if the circuit contains any other gates it raises a ValueError.
 
         Examples:
             >>> circuit = QCircuit([H(0), CNOT(0, 1), BasisMeasure()])
@@ -90,8 +95,7 @@ if InstalledProviders.QISKIT in _INSTALLED_MPQP_PROVIDERS:
         # to avoid defining twice the same parameter, we keep trace of the
         # added parameters, and we use those instead of new ones when they
         # are used more than once
-        if authorized_gates is None:
-            authorized_gates = set()
+
         qiskit_parameters = set()
         if circuit.nb_cbits == 0:
             new_circ = QuantumCircuit(circuit.nb_qubits)
@@ -109,13 +113,16 @@ if InstalledProviders.QISKIT in _INSTALLED_MPQP_PROVIDERS:
             )
 
             if isinstance(instruction, Gate):
-                from mpqp.translation.utils import verify_convert_instructions
-
-                instr = verify_convert_instructions(
-                    instruction, authorized_gates, printing
+                qiskit_gate_set = get_qiskit_gate_set()
+                instr = list(
+                    resolve_gate(
+                        instruction,
+                        qiskit_gate_set,
+                    )
                 )
             else:
                 instr = [instruction]
+
             for instruction in instr:
                 qiskit_inst = instruction.to_other_language(
                     Language.QISKIT, qiskit_parameters, **options
